refactor(client-presence): move shared workspace types to presence-definitions
Apply `@internal` tag to workspace package exports

TODO
- consider a datastore partition for common runtime structures
