+++ packages/framework/presence-states/src/latestMapValueManager.ts
@@ -28,6 +28,7 @@ import type {
 	ValidatableOptionalState,
 	ValueManager,
 } from "@fluid-internal/presence-definitions/internal";
+import type { StateDatastore } from "@fluid-internal/presence-definitions/internal/workspace-states";
 import {
 	asDeeplyReadonly,
 	asDeeplyReadonlyDeserializedJson,
@@ -38,7 +39,6 @@ import {
 	OptionalBroadcastControl,
 	toOpaqueJson,
 } from "@fluid-internal/presence-utils";
-import type { StateDatastore } from "@fluid-internal/presence-workspace";
 import { datastoreFromHandle } from "@fluid-internal/presence-workspace";
 import type { IEmitter } from "@fluidframework/core-interfaces/internal";
 import type {
