+++ packages/framework/presence-states/src/notificationsManager.ts
@@ -21,8 +21,8 @@ import type {
 	PostUpdateAction,
 	ValueManager,
 } from "@fluid-internal/presence-definitions/internal";
+import type { StateDatastore } from "@fluid-internal/presence-definitions/internal/workspace-states";
 import { brandIVM, revealOpaqueJson, toOpaqueJson } from "@fluid-internal/presence-utils";
-import type { StateDatastore } from "@fluid-internal/presence-workspace";
 import { datastoreFromHandle } from "@fluid-internal/presence-workspace";
 import type { Listeners } from "@fluidframework/core-interfaces";
 import type { JsonTypeWith } from "@fluidframework/core-interfaces/internal";
