+++ packages/framework/presence-workspace/src/index.ts
@@ -3,20 +3,10 @@
  * Licensed under the MIT License.
  */
 
-export type {
-	ClientUpdateEntry,
-	PresenceStatesInternal,
-	RuntimeLocalUpdateOptions,
-	ValueElementMap,
-} from "./presenceStates.js";
 export {
 	createPresenceStates,
 	mergeUntrackedDatastore,
 	mergeValueDirectory,
 } from "./presenceStates.js";
 
-export type {
-	LocalStateUpdateOptions,
-	StateDatastore,
-} from "./stateDatastore.js";
 export { datastoreFromHandle } from "./stateDatastore.js";
